remove secret-based sa token client builder
