+++ staging/src/k8s.io/controller-manager/go.mod
@@ -7,12 +7,14 @@ go 1.15
 require (
 	github.com/spf13/pflag v1.0.5
 	github.com/stretchr/testify v1.6.1
+	golang.org/x/oauth2 v0.0.0-20200107190931-bf48bf16ab8d
 	k8s.io/api v0.0.0
 	k8s.io/apimachinery v0.0.0
 	k8s.io/apiserver v0.0.0
 	k8s.io/client-go v0.0.0
 	k8s.io/component-base v0.0.0
 	k8s.io/klog/v2 v2.5.0
+	k8s.io/utils v0.0.0-20201110183641-67b214c5f920
 )
 
 replace (
