+++ test/integration/auth/BUILD
@@ -29,7 +29,6 @@ go_test(
         "//pkg/apis/extensions:go_default_library",
         "//pkg/apis/rbac/v1:go_default_library",
         "//pkg/auth/authorizer/abac:go_default_library",
-        "//pkg/controller:go_default_library",
         "//pkg/controller/serviceaccount:go_default_library",
         "//pkg/controlplane:go_default_library",
         "//pkg/features:go_default_library",
@@ -86,6 +85,7 @@ go_test(
         "//staging/src/k8s.io/client-go/util/keyutil:go_default_library",
         "//staging/src/k8s.io/cluster-bootstrap/token/api:go_default_library",
         "//staging/src/k8s.io/component-base/featuregate/testing:go_default_library",
+        "//staging/src/k8s.io/controller-manager/pkg/clientbuilder:go_default_library",
         "//test/e2e/lifecycle/bootstrap:go_default_library",
         "//test/integration:go_default_library",
         "//test/integration/framework:go_default_library",
